+++ src/main/java/htsjdk/samtools/util/BlockCompressedInputStream.java
@@ -689,14 +689,15 @@ public class BlockCompressedInputStream extends InputStream implements LocationA
 
     /**
      * read as many bytes as dst's capacity into dst or throw if that's not possible
+     *
      * @throws EOFException if channel has fewer bytes available than dst's capacity
      */
     static void readFully(SeekableByteChannel channel, ByteBuffer dst) throws IOException {
         int totalBytesRead = 0;
         final int capacity = dst.capacity();
-        while( totalBytesRead < capacity){
+        while (totalBytesRead < capacity) {
             final int bytesRead = channel.read(dst);
-            if( bytesRead == -1 ){
+            if (bytesRead == -1) {
                 throw new EOFException();
             }
             totalBytesRead += bytesRead;
